+++ include/kspc/math_basics.hpp
@@ -2,7 +2,6 @@
 #pragma once
 #include <array>
 #include <cmath> // sqrt, round
-#include <complex>
 #include <vector>
 #include <kspc/core.hpp>
 
@@ -118,88 +117,6 @@ namespace kspc {
   inline constexpr auto mapping_column_major(const std::size_t lda) {
     return mapping_transpose(mapping_row_major(lda));
   }
-
-  /// @}
-} // namespace kspc
-
-// projection
-namespace kspc {
-  /// @addtogroup matrix
-  /// @{
-
-  // identity function
-
-  /// %identity_fn
-  struct identity_fn {
-    using is_transparent = void;
-
-    template <typename T>
-    constexpr T&& operator()(T&& t) const noexcept {
-      return std::forward<T>(t);
-    }
-  }; // struct identity_fn
-
-  inline namespace cpo {
-    /// identity
-    inline constexpr identity_fn identity{};
-  } // namespace cpo
-
-  // consistent complex access
-
-  /// %is_complex
-  template <typename T>
-  struct is_complex : std::false_type {};
-
-  /// partial specialization of `is_complex`
-  template <typename T>
-  struct is_complex<std::complex<T>> : std::true_type {};
-
-  /// helper variable template for `is_complex`
-  template <typename T>
-  inline constexpr bool is_complex_v = is_complex<T>::value;
-
-  /// %complex_traits
-  template <typename T>
-  struct complex_traits {
-    using value_type = T;
-  };
-
-  /// partial specialization of `complex_traits`
-  template <typename T>
-  struct complex_traits<std::complex<T>> {
-    using value_type = T;
-  };
-
-  /// partial specialization of `complex_traits`
-  template <typename T>
-  struct complex_traits<const T> : complex_traits<T> {};
-
-  /// complex_value_t
-  template <typename T>
-  using complex_value_t = typename complex_traits<remove_cvref_t<T>>::value_type;
-
-  /// %conj_fn
-  struct conj_fn {
-    using is_transparent = void;
-
-    template <typename T, std::enable_if_t<is_complex_v<std::decay_t<T>>, std::nullptr_t> = nullptr>
-    constexpr auto operator()(T&& x) const noexcept(noexcept(std::conj(std::forward<T>(x))))
-      -> decltype(std::conj(std::forward<T>(x))) {
-      return std::conj(std::forward<T>(x));
-    }
-
-    // same as `identity`
-    template <typename T,
-              std::enable_if_t<!is_complex_v<std::decay_t<T>>, std::nullptr_t> = nullptr>
-    constexpr T&& operator()(T&& x) const noexcept {
-      return std::forward<T>(x);
-    }
-  }; // struct conj_fn
-
-  inline namespace cpo {
-    /// conj
-    inline constexpr conj_fn conj{};
-  } // namespace cpo
 
   /// @}
 } // namespace kspc
